--- packages/renderer/src/node/renderToStream.js
@@ -1,9 +1,0 @@
-import { pdf } from '../index';
-
-export const renderToStream = async element => {
-  const instance = pdf(element);
-  const buffer = await instance.toBuffer();
-  return buffer;
-};
-
-export default renderToStream;
